--- controllers/index.js
@@ -1,17 +1,0 @@
-const {
-  getAll,
-  getById,
-  add,
-  updateById,
-  updateStatusContact,
-  deleteById,
-} = require("./contacts-controller");
-
-module.exports = {
-  getAll,
-  getById,
-  add,
-  updateById,
-  updateStatusContact,
-  deleteById,
-};
