+++ routes/api/contacts-router.js
@@ -7,9 +7,9 @@ const {
   updateById,
   updateStatusContact,
   deleteById,
-} = require("../../controllers");
+} = require("../../controllers/contacts");
 const { validateBody, isValidId } = require("../../middlewares");
-const { schemas } = require("../../models/contact");
+const { schemas } = require("../../models");
 
 const router = express.Router();
